requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
uWSGI==2.0.21
